Make sure InputArgument ctor for int doesn't change the value

+++ source/core/Native.hpp
@@ -41,7 +41,7 @@ namespace GTA
 			InputArgument(System::UInt64 value);
 			InputArgument(System::Object ^value);
 			inline InputArgument(bool value) : InputArgument(static_cast<bool>(value) ? 1 : 0) { }
-			inline InputArgument(int value) : InputArgument(static_cast<System::UInt64>(value)) { }
+			inline InputArgument(int value) : InputArgument(static_cast<System::UInt32>(value)) { }
 			inline InputArgument(unsigned int value) : InputArgument(static_cast<System::UInt64>(value)) { }
 			inline InputArgument(float value) : InputArgument(System::BitConverter::ToUInt32(System::BitConverter::GetBytes(static_cast<float>(value)), 0)) { }
 			inline InputArgument(double value) : InputArgument(System::BitConverter::ToUInt32(System::BitConverter::GetBytes(static_cast<float>(static_cast<double>(value))), 0)) { }
